// src/interfaces/styles/formCustomStyle.ts
export interface FormCustomStyle {
  form_bg_color: string;
  form_text_color: string;
  form_bg_color_board: string;
  form_bg_color_head: string;
  form_color_text_head: string;
  form_bg_color_button: string;
  form_color_text_button: string;
  form_color_outline_text_area: string;
  form_bg_color_text_area: string;
}